LayoutTests/inspector/css/stylesheet-events-multiple-documents.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<link rel="stylesheet" href="resources/external.css">
<script>
function addSubframe() {
    var iframe = document.createElement("iframe");
    iframe.id = "subframe";
    iframe.src = "resources/stylesheet-events-subframe.html";
    document.body.appendChild(iframe);
}

function removeSubframe() {
    document.getElementById("subframe").remove();
}

function test()
{
    function sanitizeURL(url) {
        return url.replace(/^.*?LayoutTests\//, "");
    }

    function logStyleSheets() {
        WebInspector.cssStyleManager.styleSheets.sort((a, b) => a.url.localeCompare(b.url)).forEach(function(styleSheet) {
            var frameString = styleSheet.parentFrame.id === WebInspector.frameResourceManager.mainFrame.id ? "<mainframe>" : "<childframe>";
            InspectorTest.log(`URL: ${sanitizeURL(styleSheet.url)} ${frameString}`);
        });
    }

    let suite = InspectorTest.createAsyncSuite("CSS.StyleSheetEvents.MultipleDocuments");

    suite.addTestCase({
        name: "CheckStyleSheets",
        description: "Ensure there is currently one stylesheet.",
        test: (resolve, reject) => {
            InspectorTest.expectThat(WebInspector.cssStyleManager.styleSheets.length === 1, "Should be one stylesheets.");
            logStyleSheets();
            resolve();
        }
    });

    suite.addTestCase({
        name: "CheckStyleSheetsAfterAddingSubframe",
        description: "Ensure subframe stylesheets are added.",
        test: (resolve, reject) => {
            InspectorTest.evaluateInPage("addSubframe()");
            WebInspector.cssStyleManager.singleFireEventListener(WebInspector.CSSStyleManager.Event.StyleSheetAdded, function(event) {
                InspectorTest.expectThat(WebInspector.cssStyleManager.styleSheets.length === 2, "Should be two stylesheets.");
                InspectorTest.expectThat(event.data.styleSheet.parentFrame !== WebInspector.frameResourceManager.mainFrame.id, "New StyleSheet is for a subframe");
                logStyleSheets();
                resolve();
            });
        }
    });

    suite.addTestCase({
        name: "CheckStyleSheetsAfterRemovingSubframe",
        description: "Ensure subframe stylesheets are removed.",
        test: (resolve, reject) => {
            InspectorTest.evaluateInPage("removeSubframe()");
            WebInspector.cssStyleManager.singleFireEventListener(WebInspector.CSSStyleManager.Event.StyleSheetRemoved, function(event) {
                InspectorTest.expectThat(WebInspector.cssStyleManager.styleSheets.length === 1, "Should be one stylesheet.");
                InspectorTest.expectThat(event.data.styleSheet.parentFrame !== WebInspector.frameResourceManager.mainFrame.id, "Removed StyleSheet is for a subframe");
                logStyleSheets();
                resolve();
            });
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
    <p>CSS.styleSheetAdded and CSS.styleSheetRemoved for multiple documents.</p>
</body>
</html>
